thme

+++ app/assets/javascripts/events.js
@@ -19,14 +19,14 @@ function resizeEvent(event, delta) {
 function showEventDetails(event) {
 	$("#desc_dialog").find('.modal-header').show();
 	//$('#event_desc').html(event.description);
-	$('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")'>Edit</a>");
+	$('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")' class='btn btn-sm'>Edit</a>");
 	
 		title = "<h3>"+event.title+"</h3>";
 		title += "<h4>Class: "+event.description.location+"</h4>";
 		title += "<h4>Date: "+moment(event.start).format('MM/DD/YYYY')+"</h4>";		
 		title += "<h4>Timings: "+moment(event.start).format('h:mm a')+" - "+moment(event.end).format('h:mm a')+"</h4>";
 		title += "<p>"+event.description.description+"<p>";
-		$('#delete_event').html("<a href = 'javascript:void(0);' onclick ='deleteEvent(" + event.id + ", " + false + ")'>Delete</a>");
+		$('#delete_event').html("<a href = 'javascript:void(0);' onclick ='deleteEvent(" + event.id + ", " + false + ")' class='btn btn-sm'>Delete</a>");
 	
 	
 	$('#desc_dialogLabel').html(title);
@@ -37,14 +37,14 @@ function showEventDetails(event) {
 function show_list_events_details (event) {
   $("#desc_dialog").find('.modal-header').show();
 	//$('#event_desc').html(event.description);
-	$('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")'>Edit</a>");
+	$('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")' class='btn btn-sm'>Edit</a>");
 	
 		title = "<h3>"+event.title+"</h3>";
 		title += "<h4>Class: "+event.location+"</h4>";
 		title += "<h4>Date: "+moment(event.date).format('MM/DD/YYYY')+"</h4>";		
 		title += "<h4>Timings: "+moment(event.starttime).format('h:mm a')+" - "+moment(event.endtime).format('h:mm a')+"</h4>";
 		title += "<p>"+event.description+"<p>";
-		$('#delete_event').html("<a href = 'javascript:void(0);' onclick ='deleteEvent(" + event.id + ", " + false + ")'>Delete</a>");
+		$('#delete_event').html("<a href = 'javascript:void(0);' onclick ='deleteEvent(" + event.id + ", " + false + ")' class='btn btn-sm'>Delete</a>");
 	
 	
 	$('#desc_dialogLabel').html(title);
@@ -55,7 +55,7 @@ function show_list_events_details (event) {
 function showEventDetailsUser(event) {
 	$("#desc_dialog").find('.modal-header').show();
 	//$('#event_desc').html(event.description);
-	// $('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")'>Edit</a>");
+	// $('#edit_event').html("<a href = 'javascript:void(0);' onclick ='editEvent(" + event.id + ")' class='btn btn-sm'>Edit</a>");
 
 	    title = "<h3>"+event.title+"</h3>";
 		title += "<h4>Class: "+event.location+"</h4>";
@@ -78,15 +78,18 @@ function editEvent(event_id) {
 }
 
 function deleteEvent(event_id, delete_all) {
-	alert("Delete Event");
-	jQuery.ajax({
-		// data : 'authenticity_token=' + authenticity_token + '&delete_all=' + delete_all,
-		data : '&delete_all=' + delete_all,
-		dataType : 'script',
-		type : 'delete',
-		url : "/events/" + event_id,
-		success : refetch_events_and_close_dialog
-	});
+	var r = confirm("Are you Sure?");
+		if (r == true) {
+		    jQuery.ajax({
+			// data : 'authenticity_token=' + authenticity_token + '&delete_all=' + delete_all,
+			data : '&delete_all=' + delete_all,
+			dataType : 'script',
+			type : 'delete',
+			url : "/events/" + event_id,
+			success : refetch_events_and_close_dialog
+		});
+	} 
+	
 }
 
 function refetch_events_and_close_dialog() {
@@ -185,7 +188,7 @@ $(document).on("click",'#new_event',function(event) {
 			},
 			success : function(data) {
 				// $('#create_event').replaceWith(data['form']);
-				$('#eventModalLabel').html('New Event');
+				$('#eventModalLabel').html('New Schedule');
 				$('#create_event').empty();
 				$('#create_event').html(data['form']);
 				$('#create_event').find("#event_user_id").val(user_id);
